### 2025-09-15 #8919 @Kevin-Mattheus-Moerman
Please note that JOSS is currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
