+++ paper.bib
@@ -43,7 +43,8 @@
 	date = {2025},
 	publisher = {R Foundation for Statistical Computing},
 	url = {https://www.R-project.org/},
-	address = {Vienna, Austria}
+	address = {Vienna, Austria},
+	doi = {10.32614/r.manuals}
 }
 
 @techreport{motionlabsystems2021,
