### 2025-10-01 #8919 @smnnlt
@Kevin-Mattheus-Moerman Please also note that this software has already been reviewed by rOpenSci, see https://github.com/ropensci/software-review/issues/686
