### 2026-01-06 3a59a81
Update paper to new JOSS requirements
