+++ paper.md
@@ -14,7 +14,7 @@ affiliations:
  - name: Institute of Movement and Neurosciences, German Sport University Cologne, Cologne, Germany
    index: 1
    ror: 0189raq88
-date: 22 August 2025
+date: 06 January 2025
 bibliography: paper.bib
 ---
 
@@ -35,20 +35,52 @@ With `c3dr`, researchers can use R to modify, visualize, and model biomechanical
 The format is widely regarded as the standard data container in the field of human motion analysis.
 However, despite its widespread adoption, the available technical tools for working with *C3D* files are limited.
 
-The primary software for handling *C3D* files is the Biomechanical-ToolKit (`BTK`) [@barre2014].
-Released as open-source software and popular for its graphical user interface Mokka, `BTK` has long served as the standard tool for biomechanical data analysis; however, the software has not been updated since 2015.
-More recent approaches, such as the `ezc3d` library [@michaud2021], are available, but like `BTK`, they are limited to C++, MATLAB, or Python [@johnsson2022].
-
 The *R* language is widely used for scientific visualization, data wrangling, and statistical analysis [@rcoreteam2025].
 Yet, no software for reading and analyzing *C3D* files has existed in the R environment.
 We here present `c3dr`, an R package built on the open-source C++ library `ezc3d` [@michaud2021].
 The package allows users to import, modify, and export *C3D* files directly within R.
 This makes it possible to create fully reproducible data analysis pipelines for biomechanical research projects using R.
 
+# State of the field
+
+The primary software for handling *C3D* files is the Biomechanical-ToolKit (`BTK`) [@barre2014].
+Released as open-source software and popular for its graphical user interface Mokka, `BTK` has long served as the standard tool for biomechanical data analysis; however, the software has not been updated since 2015.
+More recent approaches, such as the `ezc3d` library [@michaud2021], are available, but like `BTK`, they are limited to C++, MATLAB, or Python [@johnsson2022].
+
+To work with *C3D* data in R without a dedicated package, several cumbersome options exist.
+First, the data could be exported via graphical user interfaces as a *CSV* file and then imported in R.
+However, this workflow is not reproducible, inefficient, and does not preserve the metadata of the recording.
+Second, one could use the Python interface for `ezc3d` in R via the `reticulate` package [@reticulate], but this requires expertise in both Python and R and conversion of objects between the two is suboptimal for *C3D* data.
+Third, one could use the SWIG port available for R, but a dedicated software provided as an R package is easier to install and to use.
+
 The `c3dr` package is available from CRAN and can be installed via `install.packages("c3dr")`.
 Apart from R and the `Rcpp` package for integrating the C++ source code [@eddelbuettel2011], no additional dependencies are required.
 The software is compatible with all major operating systems and contemporary C++ compilers.
 The package is released as open-source software under an MIT license.
+
+# Software Design
+
+The core of `c3dr` is the reading and writing functions for *C3D* data that wrap the existing C++ functionality.
+A key challenge was to translate the structure of *C3D* data into an appropriate structure for R objects.
+Internally, `c3dr` treats *C3D* data as nested lists with different categories.
+This guarantees that all data and metadata are imported and available for re-export.
+Because nested lists are difficult to handle in R, several accessor functions are provided to convert the list data into more accessible data (and back), such as data.frame, which can be directly used for statistical analysis or visualization.
+When writing an *C3D* file, `c3dr` recreates the file from scratch, even if no modifications were made.
+This means that importing and re-exporting a file with the package may lead to minor modifications in the structure of the file.
+But rebuilding the file based on the available data is the only way to create valid files in response to changes to the data undertaken in R and it also ensures uniform output formats regardless of the input *C3D* format.
+
+# Research Impact Statement
+
[…]
