+++ paper.bib
@@ -69,3 +69,12 @@
 	url = {https://doi.org/10.18637/jss.v040.i08},
 	langid = {en}
 }
+
+@Manual{reticulate,
+  title = {reticulate: Interface to 'Python'},
+  author = {Kevin Ushey and JJ Allaire and Yuan Tang},
+  year = {2025},
+  note = {R package version 1.44.1},
+  url = {https://rstudio.github.io/reticulate/},
+  doi = {10.32614/CRAN.package.reticulate},
+}
