### 2026-01-18 #8919 @arfon
I'll take this one as it's an rOpenSci submission.
