### 2026-01-20 #8919 @smnnlt
Hi @arfon , Thank you for taking up my submission.

@rkurchin had alerted me previously of the new JOSS submission guidelines, so I have already updated the paper to include the new sections (see thread). Let me know if further modification are needed.
