### 2026-01-28 #8919 @smnnlt
Hi @arfon

Thanks for the reminder. The current version of the article in the `paper` branch contains all four new sections. I'm looking forward to your feedback.
