### 2026-02-11 #8919 @arfon
@astamm, @japilo – adding you as reviewers here to ensure you're credited/listed on the JOSS paper too, but jus to be clear there's no additional work here to do on your part!
