### 2026-02-11 #9951 @editorialbot
[REVIEW]: c3dr: An R Package to Read and Write C3D Motion Capture Files
**Submitting author:** @smnnlt (<a href="http://orcid.org/0000-0003-1643-1860">Simon Nolte</a>)
**Repository:** https://github.com/ropensci/c3dr
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
