### 2026-02-11 #9951 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
